Moved package

+++ backend/db/migrate/20201027194327_add_pubnub_id_to_users.rb
@@ -1,5 +1,5 @@
 class AddPubnubIdToUsers < ActiveRecord::Migration[6.0]
   def change
-    add_column :users, :pubnub_id, :float
+    add_column :users, :pubnub_id, :integer
   end
 end
